+++ src/components/index.ts
@@ -1,0 +1,5 @@
+export * from "./Footer";
+export * from "./Navbar";
+export * from "./HeroSection";
+export * from "./MainLayout";
+export * from "./PageContainer";
