+++ tsconfig.json
@@ -11,14 +11,14 @@
     "esModuleInterop": true,
     "module": "esnext",
     "moduleResolution": "node",
-
     "resolveJsonModule": true,
     "isolatedModules": true,
     "jsx": "preserve",
     "baseUrl": ".",
     "paths": {
       "@/components/*": ["src/components/*"],
-      "@/images/*": ["src/image/*"]
+      "@/images/*": ["src/image/*"],
+      "@/data/*": ["src/data/*"]
     }
   },
   "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx"],
